style(design): FINDING-001 — fix nav button touch targets (44px minimum)
- Add --space-11 (44px) to CSS spacing scale
- Add space-11 to Tailwind config
- Update nav buttons to use min-h-11 for proper touch targets
- Fix mobile menu and close buttons to 44x44px

+++ backend-java/src/main/java/com/heartmirror/service/DashboardService.java
@@ -70,7 +70,7 @@ public class DashboardService {
      */
     private DashboardDTO.Overview getOverview(Long userId) {
         Long totalSessions = chatSessionRepository.countByUserId(userId);
-        Long totalDiaries = diaryRepository.findByUserIdOrderByDateDesc(userId).size();
+        Long totalDiaries = (long) diaryRepository.findByUserIdOrderByDateDesc(userId).size();
         Long totalInterventions = interventionSessionRepository.countCompletedByUserId(userId);
         Integer currentStreak = calculateStreak(userId);
